# API/backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: WarrantyVault API
  version: 1.0.0
  description: |
    WarrantyVault — a web platform for storing receipts, warranty cards,
    manuals and service history for every product you own, with automatic
    warranty-expiry tracking and in-app notifications.

    * Every authenticated endpoint requires a JWT bearer token
      (`Authorization: Bearer <token>`) obtained from `POST /api/v1/auth/login`
      or `POST /api/v1/auth/register`.
    * All responses use a consistent envelope: `{ success, message, data }`
      for 2xx and `{ success, message, errors, error }` for failures.
      `error` carries a stable `type` (validation / authentication /
      authorization / conflict / not_found / rate_limited / external /
      internal / bad_request) and the `requestId` for server-log tracing.
    * Every response includes an `X-Request-ID` header — quote it when
      reporting an error.
    * Collection endpoints are paginated with `page` / `limit` query params
      and return `{ …, pagination: { page, limit, total, totalPages } }`.

    Source: `API/backend/docs/openapi.yaml` · Human reference:
    `API/backend/docs/API-REFERENCE.md` · Interactive list: `GET /api/docs`.
...
servers:
  - url: http://localhost:5000
    description: Local development
  - url: https://warrantyvault.onrender.com
    description: Production (Render)
security:
  - bearerAuth: []
paths:
...
  /api/v1/products:
    get:
      tags: [Products]
      summary: List products (paginated, sortable, filterable)
      description: |
        Phase 4 §9/§12 — advanced server-side filters. All are optional and
        combinable: category, brand, lifecycleStatus, warrantyStatus
        (engine-derived: active/expiring_soon/expired/not_started/unknown),
        purchaseStore, repeated `tags` (AND semantics), minPrice/maxPrice,
        purchaseFrom/purchaseTo and expiryFrom/expiryTo date ranges.
      parameters:
        - { name: page, in: query, schema: { type: integer, minimum: 1, default: 1 } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
        - { name: sortBy, in: query, schema: { type: string, enum: [createdAt, updatedAt, productName, warrantyExpiryDate, purchasePrice], default: createdAt } }
        - { name: order, in: query, schema: { type: string, enum: [asc, desc], default: desc } }
        - { name: category, in: query, schema: { type: string } }
        - { name: brand, in: query, schema: { type: string } }
        - { name: lifecycleStatus, in: query, schema: { type: string, enum: [owned, in_use, stored, under_repair, sold, gifted, disposed] } }
        - { name: warrantyStatus, in: query, schema: { type: string, enum: [active, expiring_soon, expired, not_started, unknown] } }
        - { name: purchaseStore, in: query, schema: { type: string } }
        - { name: tags, in: query, schema: { type: string }, description: Repeatable; products must have all given tags (AND) }
        - { name: minPrice, in: query, schema: { type: number, minimum: 0 } }
        - { name: maxPrice, in: query, schema: { type: number, minimum: 0 } }
        - { name: purchaseFrom, in: query, schema: { type: string, format: date } }
        - { name: purchaseTo, in: query, schema: { type: string, format: date } }
        - { name: expiryFrom, in: query, schema: { type: string, format: date } }
        - { name: expiryTo, in: query, schema: { type: string, format: date } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  message: { type: string }
                  data: { $ref: "#/components/schemas/ProductList" }
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [Products]
      summary: Create a product
      description: |
        Serial numbers are unique per user among active products
        (DB-enforced); a duplicate returns 409.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProductInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  message: { type: string }
                  data: { $ref: "#/components/schemas/Product" }
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationError"
...
  /api/v1/export/products:
    get:
      tags: [Products]
      summary: Export all owned products (JSON, CSV or ODS)
      description: |
        Phase 4 §16 — downloads every live product with its service history.
        Ownership is scoped to the authenticated user. CSV follows RFC 4180
        (quoting/escaping); ODS is an OpenDocument Spreadsheet with a bold
        header row. Delivered as an attachment.
      parameters:
        - { name: format, in: query, schema: { type: string, enum: [json, csv, ods], default: json } }
      responses:
        "200":
          description: File attachment (application/json, text/csv or application/vnd.oasis.opendocument.spreadsheet)
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /api/v1/products/{productId}/shares:
    parameters:
      - { name: productId, in: path, required: true, schema: { type: string, description: Mongo ObjectId } }
    get:
      tags: [Products]
      summary: List the owner's share links for a product (active + revoked)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  message: { type: string }
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/Share" }
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      tags: [Products]
      summary: Create a read-only share link for a product
      description: |
        Phase 4 §17 — grants read-only access to exactly one product via an
        unguessable 48-hex-char token. Optional expiry 1–90 days (omitted =
        never expires, still revocable).
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                expiresInDays: { type: integer, minimum: 1, maximum: 90, nullable: true }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  message: { type: string }
                  data: { $ref: "#/components/schemas/Share" }
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/ValidationError"
...
  /api/v1/products/{productId}/documents:
    parameters:
      - { name: productId, in: path, required: true, schema: { type: string, description: Mongo ObjectId } }
    get:
      tags: [Documents]
      summary: List documents for a product (paginated)
      parameters:
        - { name: page, in: query, schema: { type: integer, minimum: 1, default: 1 } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 100 } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  message: { type: string }
                  data: { $ref: "#/components/schemas/DocumentList" }
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      tags: [Documents]
      summary: Upload a document for a product
      description: |
        Rate-limited (10 per minute per user). Eligible types (receipt,
        warranty_card) are OCR'd asynchronously; PDFs get their text layer
        read (scanned pages are rasterized, max 3 pages).
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [documentType, file]
              properties:
                documentType:
                  type: string
                  enum: [receipt, warranty_card, product_photo, manual, other]
                notes: { type: string, maxLength: 2000 }
                file:
                  type: string
                  format: binary
                  description: JPEG, PNG, WEBP or PDF, max 5 MB (magic bytes verified)
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  message: { type: string }
                  data: { $ref: "#/components/schemas/Document" }
        "400":
          $ref: "#/components/responses/BadRequest"
        "422":
          $ref: "#/components/responses/ValidationError"
        "429":
          $ref: "#/components/responses/RateLimited"
...
  /api/v1/documents:
    get:
      tags: [Documents]
      summary: List all of the user's documents (paginated)
      parameters:
        - { name: page, in: query, schema: { type: integer, minimum: 1, default: 1 } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 100 } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  message: { type: string }
                  data: { $ref: "#/components/schemas/DocumentList" }
    post:
      tags: [Documents]
      summary: Upload a standalone document (no product yet)
      description: Same rules as the product-scoped upload; OCR runs the same way.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [documentType, file]
              properties:
                documentType:
                  type: string
                  enum: [receipt, warranty_card, product_photo, manual, other]
                notes: { type: string, maxLength: 2000 }
                file: { type: string, format: binary, description: JPEG, PNG, WEBP or PDF, max 5 MB }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  message: { type: string }
                  data: { $ref: "#/components/schemas/Document" }
        "429":
          $ref: "#/components/responses/RateLimited"
...
  /api/v1/products/{productId}/service-history:
    parameters:
      - { name: productId, in: path, required: true, schema: { type: string } }
    get:
      tags: [Service History]
      summary: List service records for a product (capped at 200, newest first)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  message: { type: string }
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/ServiceRecord" }
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags: [Service History]
      summary: Add a service record
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ServiceRecordInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  message: { type: string }
                  data: { $ref: "#/components/schemas/ServiceRecord" }
        "422":
          $ref: "#/components/responses/ValidationError"
...
  /api/v1/notifications:
    get:
      tags: [Notifications]
      summary: List notifications (paginated)
      parameters:
        - { name: unreadOnly, in: query, schema: { type: boolean, default: false } }
        - { name: page, in: query, schema: { type: integer, minimum: 1, default: 1 } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 50 } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  message: { type: string }
                  data: { $ref: "#/components/schemas/NotificationList" }
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /api/v1/migrations:
    post:
      tags: [Migration]
      summary: One-step upload and import
      description: Ingests, normalizes, validates, deduplicates, and creates products in a single operation.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
          application/json:
            schema:
              type: object
              properties:
                content: { type: string }
                fileName: { type: string }
      responses:
        "201":
          description: Migration completed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  message: { type: string, example: Migration completed successfully }
                  data: { $ref: "#/components/schemas/MigrationDetail" }
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
    get:
      tags: [Migration]
      summary: List user's past migrations
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
      responses:
        "200":
          description: Paginated list of migrations
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  message: { type: string }
                  data:
                    type: object
                    properties:
                      migrations:
                        type: array
                        items: { $ref: "#/components/schemas/MigrationSummary" }
                      total: { type: integer }
                      page: { type: integer }
                      limit: { type: integer }
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token from POST /api/v1/auth/login or /register.
  responses:
...
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: Missing or invalid token
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: Authenticated but not the owner of the resource
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Conflict:
      description: Duplicate resource (e.g. serial number already used)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    ValidationError:
      description: Validation failed (422)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    RateLimited:
      description: Too many requests (429)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
...
    Pagination:
      type: object
      properties:
        page: { type: integer, example: 1 }
        limit: { type: integer, example: 20 }
        total: { type: integer, example: 137 }
        totalPages: { type: integer, example: 7 }
...
    Product:
      type: object
      properties:
        _id: { type: string }
        userId: { type: string, description: Owner ObjectId }
        productName: { type: string }
        brand: { type: string }
        model: { type: string }
        category: { type: string }
        purchaseDate: { type: string, format: date, nullable: true }
        purchasePrice: { type: number, minimum: 0, nullable: true }
        currency: { type: string, nullable: true }
        purchaseStore: { type: string, nullable: true }
        serialNumber: { type: string, nullable: true }
        warrantyExpiryDate: { type: string, format: date, nullable: true }
        warrantyPeriodMonths: { type: integer, nullable: true }
        warrantyStatus: { type: string, enum: [not_started, active, expiring_soon, expired, unknown], description: Primary warranty status computed by the canonical engine (Phase 4 §5) }
        warrantyStatusLabel: { type: string }
        warrantyProvider: { type: string, nullable: true }
        warrantyProviderType: { type: string, enum: [manufacturer, retailer, third_party, extended, unknown], nullable: true }
        warrantyContact: { type: string, nullable: true }
        warrantyWebsite: { type: string, nullable: true }
        lifecycleStatus: { type: string, enum: [owned, in_use, stored, under_repair, sold, gifted, disposed], default: owned }
        warranties: { type: array, items: { $ref: "#/components/schemas/WarrantyPeriod" }, default: [] }
        tags:
          type: array
          description: User-scoped tags, lower-cased + trimmed (Phase 4 §12)
          items: { type: string }
        notes: { type: string, nullable: true }
        isDeleted: { type: boolean }
        thumbnailUrl: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    ProductInput:
      type: object
      properties:
        productName: { type: string, minLength: 1 }
        brand: { type: string, maxLength: 100 }
        model: { type: string, maxLength: 100 }
        category: { type: string, maxLength: 100 }
        purchaseDate: { type: string, format: date }
        purchasePrice: { type: number, minimum: 0 }
        currency: { type: string, maxLength: 10 }
        purchaseStore: { type: string, maxLength: 200 }
        serialNumber: { type: string, maxLength: 200 }
        warrantyExpiryDate: { type: string, format: date }
        warrantyPeriodMonths: { type: integer, minimum: 1 }
        warrantyProvider: { type: string, maxLength: 200 }
        warrantyProviderType: { type: string, enum: [manufacturer, retailer, third_party, extended, unknown] }
        warrantyContact: { type: string, maxLength: 200 }
        warrantyWebsite: { type: string, format: uri }
        lifecycleStatus: { type: string, enum: [owned, in_use, stored, under_repair, sold, gifted, disposed] }
        warranties: { type: array, maxItems: 20, items: { $ref: "#/components/schemas/WarrantyPeriodInput" } }
        tags:
          type: array
          description: User-scoped tags (max 20, each ≤ 30 chars, lower-cased server-side)
          maxItems: 20
          items: { type: string, maxLength: 30 }
        notes: { type: string, maxLength: 2000 }
        thumbnailUrl: { type: string, format: uri }
      description: productName is required on create; warrantyExpiryDate must be after purchaseDate; each warranty period expiry must be after its start.
...
    ProductList:
      type: object
      properties:
        products:
          type: array
          items: { $ref: "#/components/schemas/Product" }
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }
    Document:
      type: object
      properties:
        _id: { type: string }
        productId: { type: string, nullable: true, description: null for standalone uploads }
        userId: { type: string }
        documentType: { type: string, enum: [receipt, warranty_card, product_photo, manual, other] }
        fileName: { type: string }
        fileUrl: { type: string }
        publicId: { type: string, description: Cloudinary public id }
        fileSize: { type: integer }
        mimeType: { type: string }
        uploadedAt: { type: string, format: date-time }
        docState: { type: string, enum: [unreviewed, reviewed, important, archived], description: Phase 4 §13 organization state }
        verified: { type: boolean, description: Manual user verification — never set by OCR }
        tags: { type: array, items: { type: string }, description: User-scoped, normalized (trimmed/lowercased/deduped, max 20) }
        notes: { type: string, nullable: true }
        ocrStatus: { type: string, enum: [pending, processing, done, failed, skipped] }
        ocrText: { type: string, nullable: true }
        parsedData:
          type: object
          nullable: true
          properties:
            warrantyExpiryDate: { type: string, format: date, nullable: true }
            purchasePrice: { type: number, nullable: true }
            serialNumber: { type: string, nullable: true }
            purchaseStore: { type: string, nullable: true }
            purchaseDate: { type: string, format: date, nullable: true }
            productName: { type: string, nullable: true }
            brand: { type: string, nullable: true }
            model: { type: string, nullable: true }
        ocrError: { type: string, nullable: true }
    DocumentList:
      type: object
      properties:
        documents:
          type: array
          items: { $ref: "#/components/schemas/Document" }
        pagination: { $ref: "#/components/schemas/Pagination" }
    Share:
      type: object
      properties:
        shareId: { type: string }
        token: { type: string, description: 48 hex chars — unguessable }
        createdAt: { type: string, format: date-time }
        expiresAt: { type: string, format: date-time, nullable: true, description: null = never expires }
        revokedAt: { type: string, format: date-time, nullable: true }
        active: { type: boolean, description: false once revoked or expired }
        url: { type: string, example: /shared.html?t=<token> }
...
    ServiceRecord:
      type: object
      properties:
        _id: { type: string }
        productId: { type: string }
        userId: { type: string }
        serviceDate: { type: string, format: date }
        serviceType: { type: string, enum: [repair, maintenance, inspection, replacement, other] }
        serviceProvider: { type: string, nullable: true }
        cost: { type: number, minimum: 0, nullable: true }
        currency: { type: string, nullable: true }
        description: { type: string, nullable: true }
        documentIds:
          type: array
          items: { type: string }
        nextServiceDate: { type: string, format: date, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    ServiceRecordInput:
      type: object
      required: [serviceDate, serviceType]
      properties:
        serviceDate: { type: string, format: date }
        serviceType: { type: string, enum: [repair, maintenance, inspection, replacement, other] }
        serviceProvider: { type: string, maxLength: 200 }
        cost: { type: number, minimum: 0 }
        currency: { type: string, maxLength: 10 }
        description: { type: string, maxLength: 2000 }
        documentIds:
          type: array
          items: { type: string }
        nextServiceDate: { type: string, format: date }
...
    NotificationList:
      type: object
      properties:
        notifications:
          type: array
          items: { $ref: "#/components/schemas/Notification" }
        pagination: { $ref: "#/components/schemas/Pagination" }
...